api/requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.1
flask==2.2.3
flask-cors==3.0.10
google-api-python-client==2.107.0